backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg[binary]>=3.1,<4.0

# Geospatial (Linux)
fiona>=1.9,<2.0
shapely>=2.0,<3.0
pyproj>=3.6,<4.0

# Image processing
Pillow>=10.0,<11.0

# Environment
python-dotenv>=1.0,<2.0

# Email validation
email-validator>=2.0,<3.0
django-disposable-email-checker>=2.0,<3.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Production
gunicorn>=21.0,<23.0
whitenoise>=6.6,<7.0
